requirements.txt
streamlit
praw
python-dotenv
groq
fpdf2
requests
